requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
gunicorn==20.0.4
hiredis==1.1.0
idna==2.10
redis==3.5.3
redisearch==2.0.0
requests==2.25.1
rmtest==0.7.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
